+++ affichage/deuxD/vue2DRep/__init__.py
@@ -1,1 +1,5 @@
 from .Vue2D import *
+from .Vue2DVecteur import *
+from .Vue2DPave import *
+from .Vue2DRobot import *
+from .Vue2DArene import *
