+++ test/testAppRobot.py
@@ -5,8 +5,10 @@ Creation et affichage d'un robot basique avec modulateurs de vitesses pour teste
 """
 
 rt=Robot(Pave(50,50,0),Objet3D(),Objet3D(), Vecteur(0,-1,0))
-rt.deplacer(Vecteur(100,100,0))
+rt.deplacer(Vecteur(200,100,0))
 a=Arene()
 a.add(rt)
 
-AppRobot(rt, Vue2DArene(a)).mainloop()
+app=AppRobot(rt, Vue2DArene(a))
+app.init()
+app.mainloop()
